mod: 现在生成一个固定的模型可以展示了
